Prevent PayPal SDK expection parsing bug
PayPal SDK ServerError exception class crashes
if you ask for a message or inspect. Catch the exception
and instead log a general error msg precenting the crash.
The crash is because it always tries to parse the response
body as json and sometimes world is not entirely made of json.