Remove unnecessary PayPal permissions
Some paypal permissions that we used to request are causing issues for
customers in some specific countries. We know at least Chile and
Bulgaria to be such countries. Moreover, it turns out we probably don't
even need these permissions for anything. The bonus effect is that the
permissions list looks less intimidating for users signing up.

The permissions we remove are detailed here:
https://developer.paypal.com/docs/classic/permissions-service/integration-guide/PermissionsAbout/