Recognize nil order_ids for auth flow events
When we switched from the order flow to authorization flow authorization
related IPN messages no longer have parent_txn_id set because they don't
have a parent transaction. The parent transaction used to be the id of
the order. IPN sends this field non the less but with empty string
value. We didn't recognize the empty field to mean nil which caused an
attempt to update an incorrect payment at later stage of handling the
IPN message. Database prevents this with unique id constraint but this
causes an unnecessary error to pop up in airbrake. Also this would mean
we lose data if the web transaction flow would fail in the middle of
receiving success response from PayPal server. Properly handled IPN
message will reconcilate the state of the payment.