Show quantity * unit price as the sum in admin transaction list view
Previously for some reason (probably because of preauth) the sum that we show in the transaction list view was the total payment sum done in the payment gateway. Thus, transactions that do not use payment gateway didn't show any price. The easiest (and in my opinion also the correct way) to calculate and show the price is simply multiply quantity and unit price