Call `in_time_zone` to convert Time to TimeWithZone
From ActiveSupport CHANGELOG:

Modify TimeWithZone#as_json to include 3 decimal places of sub-second accuracy by default, which is optional as per the ISO8601 spec, but extremely useful. Also the default behaviour of Date#toJSON() in recent versions of Chrome, Safari and Firefox.

So TimeWithZone is serialized with 3 decimals for milliseconds. However, Time class doesn't have this property. So that's why in tests the actual value and the expected value both have to be TimeWithZone