Fix bug in SQL query generation
In the Search API response, the total number is separate from the count
of listings in the data. For example, if the user requests &page=2 from
a search that has less listings that fit on one page, the total count is
the number of listings in the first page (i.e. total count), but the
data array is empty, making the SQL query syntax invalid with an empty
array of ids.

The fix is to check the length of the data array to check if we need to
query the DB.