Fix broken fonts
Problem: The font-awesome fonts are currently broken after customizing the marketplace appearance. The CSS compilation creates a relative path that expects that the font is found in S3 (where the compiled CSS is saved). However, we shouldn't rely that the fonts are there.

Solution: There's no reason why the fonts should be included in the main stylesheet. We can include the font stylesheet in separate link-tag.