Move app-bundle CSS from head to body
Originally, we decided to add app-bundle CSS to the head, because we though that loading it in the body would trigger additional browser redraw and we wanted to avoid that. However, as Google Page Speed points out, adding the app-bundle in the head prevents any rendering from happening, before the app-bundle CSS is loaded. My hunch is that blocking rendering is bigger bad than triggering page redraw.