Cache access to url_helpers module in view utils
Rails.application.routes.url_helpers is ~500 times slower than just
Rails.application.routes. This is because .url_helpers dynamically
builds an involved module on every invocation. This module is intended
to be mixed in to classes in which case it gets cached as part of the
Class definition. When accessed directly in other modules it must always
be cached to avoid the costly reconstruction on every call.