Add admin controls for toggling new searchpage
Adds admin controls for searhpage. New searchpage requires topbar and
it is enabled automatically if searchpage is enabled. JS is used to
disable and enable topbar controls when searchpage is checked and
unchecked.

Adds also a rule for private marketplaces as these cannot have
searchpage if custom landing page isn't enabled.