Fix POST body proxying to Harmony in production env
The request.body is a StringIO in development environment, but a
PhusionPassenger::Utils::TeeInput in production. Calling the
request.raw_post didn't work with the TeeInput and was changed to
using request.body.read which works in both envs. However, it should
be only read once to avoid the Transit decoder draining leaving it
empty for the proxy.