Fix an issue that caused map marker clusters to disappers when zoomed
This commit fixes an issue that caused the map marker cluster to disapper when zoomed enough. The reason for this is that we have set 17 as a max zoom level. However, when the icon cluster is clicked, the max zoom level is not taken into account. The map clustering code will try to zoom even if the max zoom level is reached, and the result is that the cluster disappears but the map doesn't zoom.

Solution: Check if we can zoom. If not, show info bubble.