Fix an issue which caused super admins to see error message in map view
The code was checking only if the user is a member of the marketplace. However, since the super admins are not members of the marketplaces, they were not allows to see the listings.

Also, a deprecation comment is added to the Listing model. Authorization logic should be moved to controllers.