Map icons from bundled fontawesome to main UI
The default iconset used by the sharetribe site is Symbolset Pika,
which is proprietary. Whereas sharetribe comes bundled with
fontawesome + mappings of some actions to fontawesome.

Problem is, a lot of icons in SS Pika are not mapped to an equivalent
fontawesome icon.

This commit adds mappings for SS Pika icons into the bundled
fontawesome.