Build a web image suitable for development
This also fixes the volume to point to `/opt/app` instead. I was
modifying the db/seeds.rb, which was accessible from the container at
`opt/sharetribe` but the app was running the on `/opt/app`. As a result,
none of my changes where being picked up. I was seeing ghosts.