Fix license formatting issue
I noticed a formatting error in the license text, where a following
paragraph was enumerated in a list, rather than the final element.

I am NOT reviewing or approving the legal substance of the license.  Do
not rely on this patch for legal advice.  See https://notlegaladvice.law
for more information.