Enable customizable footer
For this to work it is necessary to enable Sharetribe's external plan
service config.

This makes the app build the `@current_plan` reading from the
`marketplace_plans` DB table (done in
lib/rack_middleware/marketplace_lookup.rb) and thus, having an instance
of `PlanService::Store::Plan::PlanModel` where features like `:footer`
are read from.

In contrast with what is mentioned in
https://github.com/sharetribe/sharetribe/issues/3740#issuecomment-487916936,
this feature depends on the plan instance (is read from the
`marketplace_plans` record), and has nothing do with feature flags.
https://github.com/coopdevs/sharetribe/blob/master/app/presenters/footer_presenter.rb#L10-L13
is pretty self explanatory. The whole feature was introduced in
https://github.com/sharetribe/sharetribe/pull/3374.

While the migration is needed to enable it in next.donalo.org, the seed
data ensures any new server, whatever RAILS_ENV it may be, has it
enabled.