Setup RSpec in Donalo engine
This moves from the default test/ to spec/ and configures RSpec to run
future engine tests as outlined in
https://corlewsolutions.com/articles/article-126-how-to-setup-rspec-in-a-rails-engine,
https://hocnest.com/blog/testing-an-engine-with-rspec/ and
https://relishapp.com/rspec/rspec-rails/docs/gettingstarted.

Unfortunately, since engine specs need to be executed from within the
engine directory, there's a bit defensive coding we need to add so that
not found constants don't break the tests. We could investigate other
potential solutions though.

We stick to SQLite3 just so we get this out the door ASAP. We're on
a very tight schedule and there are no specs touching DB yet. I followed
the make it just work and touch as little as possible, approach.

To run them, you can do:

```
$ docker-compose exec web bash
cd engines/donalo
bundle exec rake rspec/
```