Remove tests-related defensive coding
The order in which the initializer blocks are called doesn't seem to be
deterministic and now, `Listing` is not yet loaded when the
"donalo/monkey_patch/stock_control" is called thus, the stock feature is
never loaded.

We'll need to see how this affects the tests and find a better way to
distinguish between production and test.