Use the appropriate paypal and stripe services to fetch account data
The Rails model for PayPal account and especially the relationship from Person
is not reliable, as multiple PayPal account rows may exist for one person
in different configurations. The PaypalService is the authority.

For Stripe, the model technically works, but semantically it is better again
to use the appropriate service.